fix: F5 새로고침 시 불필요한 로딩페이지 제거
