+++ src/components/Dashboard/index.ts
@@ -1,0 +1,4 @@
+export { default as UserProfile } from './UserProfile';
+export { default as SkinScoreChart } from './SkinScoreChart';
+export { default as RecentDiagnosisRecords } from './RecentDiagnosisRecords';
+export { default as SkinKnowledgeCard } from './SkinKnowledgeCard';
