+++ src/constants/diseaseAnalysis.ts
@@ -1,0 +1,18 @@
+// 증상 목록
+export const SYMPTOMS = [
+  '가려움', 
+  '따가움/통증', 
+  '붉은 반점', 
+  '각질/비늘', 
+  '진물/수포', 
+  '피부 건조', 
+  '뾰루지/여드름'
+];
+
+// 기간 목록
+export const DURATIONS = [
+  '오늘', 
+  '2-3일 전', 
+  '1주일 이상', 
+  '오래 전'
+];
